POStagging/requirement.txt
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
executing==0.8.3
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.22.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
stack-data==0.2.0
tdqm==0.0.1
tqdm==4.64.0
traitlets==5.1.1
wcwidth==0.2.5
